Move ResyncCallback in `object IOPlatform`
